refactor(lsm): lru list
